requirements.txt
Django~=3.2.9
django-environ~=0.8.1
python-slugify~=5.0.2
